+++ crypto/src/dap_sign.c
@@ -466,13 +466,13 @@ int dap_sign_verify(dap_sign_t *a_chain_sign, const void *a_data, const size_t a
  * @param a_chain_sign dap_sign_t object
  * @return size_t 
  */
-size_t dap_sign_get_size(dap_sign_t * a_chain_sign)
+uint64_t dap_sign_get_size(dap_sign_t * a_chain_sign)
 {
     if (!a_chain_sign || a_chain_sign->header.type.type == SIG_TYPE_NULL) {
         debug_if(s_dap_sign_debug_more, L_WARNING, "Sanity check error in dap_sign_get_size");
         return 0;
     }
-    return (sizeof(dap_sign_t) + a_chain_sign->header.sign_size + a_chain_sign->header.sign_pkey_size);
+    return (uint64_t)sizeof(dap_sign_t) + a_chain_sign->header.sign_size + a_chain_sign->header.sign_pkey_size;
 }
 
 dap_sign_t **dap_sign_get_unique_signs(void *a_data, size_t a_data_size, size_t *a_signs_count)
